Update main.yml
Nice work! tada You added a workflow!

Here's what it means:

    name: A workflow for my Hello World file gives your workflow a name. This name appears on any pull request or in the Actions tab. The name is especially useful when there are multiple workflows in your repository.
    on: push indicates that your workflow will execute anytime code is pushed to your repository, using the push event.

Next, we need to specify a job or jobs to run.
Actions

Workflows piece together jobs, and jobs piece together steps. We'll now create a job that runs an action. Actions can be used from within the same repository, from any other public repository, or from a published Docker container image. We'll use an action that we'll define in this repository.

We'll add the block now, and break it down in the next step.
Step 5: Run an action from your workflow file

Let's add the expected action to the workflow.
keyboard Activity: Add an action block to your workflow file

    As a part of this branch and pull request, edit .github/workflows/main.yml to append the following content: